AU-2354: feature: AU-2354: Moved the hideSliePopup function to helpers.ts.
