+++ e2e/tests/forms/private_person_48.ts
@@ -9,7 +9,7 @@ import {
   fillInputField,
   fillHakijanTiedotPrivatePerson,
   uploadFile,
-  hideSlidePopup, fillFormField
+   fillFormField
 } from "../../utils/form_helpers";
 
 import {
