+++ e2e/tests/forms/registered_community_58.ts
@@ -8,7 +8,7 @@ import {
 import {
   fillFormField,
   fillGrantsFormPage, fillHakijanTiedotRegisteredCommunity, fillInputField,
-  hideSlidePopup, uploadFile
+   uploadFile
 } from '../../utils/form_helpers';
 
 import {
