+++ e2e/tests/forms/registered_community_65.ts
@@ -7,7 +7,7 @@ import {
 import {
   fillFormField,
   fillGrantsFormPage, fillHakijanTiedotRegisteredCommunity, fillInputField,
-  hideSlidePopup, uploadFile
+   uploadFile
 } from '../../utils/form_helpers';
 
 import {
