+++ e2e/tests/forms/unregistered_community_64.ts
@@ -7,7 +7,7 @@ import {
 import {
   fillFormField,
   fillGrantsFormPage, fillHakijanTiedotUnregisteredCommunity, fillInputField,
-  hideSlidePopup, uploadFile
+   uploadFile
 } from '../../utils/form_helpers';
 
 import {
